server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy",
    "psycopg2-binary",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "email-validator",
    "bcrypt==4.3.0",
]

[tool.setuptools]
packages = ["app"]
